refactor(Flight): Change startingPrice from Double to Integer and Math.ceil()

+++ backend/src/main/java/com/holstandreas/srv/model/Flight.java
@@ -53,6 +53,6 @@ public class Flight {
 
   @NotNull
   @Min(0)
-  private Double startingPrice;
+  private Integer startingPrice;
 
 }
